Implement basic framework to register land
Implemented the basic functions required to register a piece of
land. The modifiers "onlyInspector", and "onlyRegisteredUser" are
currently always passing. These modifiers will become functional once
the contract dealing with accounts is available for use.

Certain fields like a field indicating whether a land is encumbered
have been omitted for now and will be added on later.