Added basic modifiers and functions
addLandInspector allows the deployer to add a new land inspector to the system. It assigns a new UserAccount struct to the _inspector address with the provided username and sets the isLandInspector flag to true, indicating that this address is a land inspector.
removeLandInspector allows the deployer to remove land inspector privileges from an existing land inspector. It checks if the address being removed is not the deployer's address (deployer's privileges cannot be removed). If not, it assigns an empty username and sets the isLandInspector flag to false for the _inspector address, effectively revoking their land inspector status.
isLandInspector allows anyone to check whether a given Ethereum address is a land inspector.