Replace Variables with a LandRecord Structure
Replaced the variables passed to two functions with a
LandRecord struct instead due to the limit EVM sets on the
number of local variables allowed per function.
EVM allows a maximum of 16 local variables per function.

Even after replacing it with a structure, the problem still
persists.
Other things I tried was to disable yul optimization. This didn't
help either.