Update LandMapToId and add events
Updated the LandMapToId mapping. Previously the mapping was missing
the survey number and subdivision mappings. This information is necessary
to uniquely identify a piece of land. These mappings have now been added.

Added the LandRecordAdded and LandRecordExists events.
Since the addLandRecord when invoked leads to a transaction, the return
value can't be captured by external systems. Hence, these need to be
emitted as events.

When trying to add an existing land record, the LandRecordExists event
is emitted with the landId.

When a new land record has been successfully added, the LandRecordAdded
event is emitted with the newly added landId.