Add getPendingVerification, requestVerification and update verifyAccount
Added a PendingVerification struct to store pending verifications, including the user's address and their Aadhar number.
Added a getPendingVerifications function, which allows anyone to view the list of pending verifications.
In the verifyAccount function, modified the code to remove the account from the list of pending verifications when it is verified
Added a requestVerifications function, which allows user to request for verification