Remove onlyRegisteredUser modifier temporarily
Removed the onlyRegisteredUser modifier on the addLandRecords function
temporarily. There seems to be a problem with the verifyAccounts function
in the accounts smart contract. Even if a user is verified, their status
remains unverified, which can be observed by the isUserVerified functions
response.
Hence, this modifier is temporarily removed so as to continue the work on
the load testing script and shall be added later.