Add AccountRegistration load testing script
Implemented a script to load test two functions in the
AccountRegistration contract:
1. setUserDetails
2. requestVerification

The script assumes there are 1000 user accounts. It sends out concurrent
requests in batches of 50, waits for them to be resolved and repeats the
same again, for a total of 1000 or no of user accounts times.
The script logs the time required for each of these functions to execute
all the 1000 transactions.
Note: The contract addresses in the script need to be replaced each time
the script is run.