Fix landVerificationRequest and verifyLand
Previously, the landVerificationRequest would add requests to the
verificationRequired list without any checks. This could lead to
problems like having multiple verification requests for the same
land id/land record.
The verifyLand function didn't remove the requests from the
verificationRequired list once the land record is verified. Therefore,
there was no way to fetch pending land verification requests.
This commit addresses these issues.

The landVerificationRequest function now emits the event
"LandAlreadyVerified" if the land record is already verified.
It emits the event "LandVerificationRequestExists" if a verification request
for the land id already exists.
It emits the event "LandVerificationRequestSubmitted" when the
verification request is submitted.

The verifyLand function now removes the landId from the verificationRequired
list upon verification of the land record.
It also emits the "LandVerified" event upon verification of the land record.